requirements.txt
django
djangorestframework
scipy
scikit-learn
numpy
pandas